Add support for GitHub token
A GitHub token is needed due to the rate limiting of GitHub. With the
token we are able to get changelogs and other information from GitHub
and add them to the MR's description.

See: https://docs.renovatebot.com/install-gitlab-app/#configuring-a-token-for-githubcom-hosted-release-notes

We don't need to encrypt our tokne, as we run renovate ourselves and
trust the infrastructure.