Regenerate NOTICE on dependency updates in Gitaly
Gitaly's NOTICE file contains the licenses of the dependencies. This
file is verified to be up to date in Gitaly's pipeline. Updating a
dependency may require updating the NOTICE file. This commit adds a
step to update the NOTICE file after updating the dependencies. Since
the NOTICE generation job is executed via make in Gitaly, make is
added to the Docker image.