Include NOTICE file in the commit after updating dependency
In 8379b558 we added a post upgrade task to regenerate the NOTICE
file after updating dependencies. We didn't however specify the file
in the fileFilters field which defines which files end up in the commit
along with the dependency updates. This commit adds the missing fileFilter
so the generated NOTICE file is committed as well.