Generate NOTICE per dependency
Renovate opens a single MR for each updated dependency even if there
are multiple dependency updates that could be made. This means that
the executionMode of the post upgrade task should actually be per
dependency to ensure the NOTICE file only includes changes from the
dependency being updated in a given MR. This commit removes the
branch execution mode so the setting defaults back to the update.
This way each dependency change gets its own NOTICE update.