gitaly: Automatically update build tools
Gitaly has recently refactored the way it keeps track of its build
tooling to track versions in separate Go modules. Part of the motivation
was to not have to worry about upgrading thase versions manually anymore
given that we can now use standard Go tooling to do so. This means that
the Renovate bot can do this tedious work for us.

Adapt the configuration to include the `tools/` directory to enable
this.