gitaly: Revert to use `gomodTidy` instead of `gomodTidy1.17`
With cd86529 (Add gomodTidy1.17 to postUpdateOptions for Gitaly,
2022-10-19), we have switched tho `postUpdateOptions` to use
`gomodTidy1.17` instead of `gomodTidy`. Whereas `go mod tidy` will by
default behave as if the `-compat` flag was set to the version prior to
the one indicated by the `go` directive in the project's `go.mod` file,
with `gomodTidy1.17` we explicitly overrode this to drop compatibility
with Go 1.16.

This is causing breakage now though as Gitaly has dropped support for Go
1.17. So let's again revert to use `gomodTidy` so that Go figures out
automatically which versions to retain compatibility with. This both
decreases our maintenance burden and should unbreak our builds.

Changelog: fixed