Add gomodUpdateImportPaths to Gitaly options
When updating to a new major version of a dependency we need to
explicitly use the new version in our import paths. If imports are
not updated, then future invocations of `go mod tidy` will remove
the newly updated packages. Gitaly's `verify`
job is catching this issue, leading to
failed pipelines on MRs from Renovate Bot.

To resolve this, enable the `gomodUpdateImportPaths` option, which will
automatically update import paths to the new version.