Create new containerbase based image
This adds a new Dockerimage which is based on renovates binarySource:
'install' technique. This way we do not need to install other runtimes,
but node. renovate will install runtimes automatically based on the
project preferences.

This switches the gitaly project, as it has both ruby and go and custom
Make commands.

See:
- https://gitlab.com/gitlab-org/frontend/renovate-gitlab-bot/-/issues/33
- https://docs.renovatebot.com/self-hosted-configuration/#binarysource