Skip pre- and post-processing if branches do not match
Since we use dynamic child pipelines, projects like GitLab with multiple
renovate configs are processed in parallel. This means that the pre- and
post-processing of e.g. node and ruby configs run at the same time,
leading e.g. to multiple comments posted at the same time.

This introduces a guard for the unique branch prefixes of each config
which stops that behavior.