Enable immediate Gitaly MR creation
Gitaly's pipeline does not run when a non-default branch is pushed. An
MR must be created for a pipeline to be triggered. Our current
`prCreation` setting of `not-pending` requires that a pipeline succeed
before an MR is created. In the past the `renovate/stability-days` job
was treated as a successful pipeline[1] and would trigger MR creation and
a full pipeline run[2].

However, Renovate-Bot v35 added a new `internalChecksAsSuccess`
option[3] that defaults to requiring a non-renovate pipeline to have
succeeded before creating an MR. As a result, with 978e7d7 (Update
renovate to v35.79, 2023-10-10), all new branches on the Gitaly fork were
treated as being in 'pending' status and never had an MR created.

As Gitaly will not run a pipeline without an MR, set `prCreation` to
`immediate` and `internalChecksFilter` to `strict`. This will suppress
creation of a branch[4] until the new release has reached its age
criteria, at which point the branch and MR are created immediately.

[1] https://gitlab.com/gitlab-renovate-forks/gitaly/-/pipelines/794330788
[2] https://gitlab.com/gitlab-renovate-forks/gitaly/-/pipelines/794330892
[3] https://github.com/renovatebot/renovate/pull/20572
[4] https://docs.renovatebot.com/configuration-options/#suppress-branchpr-creation-for-x-days