Wait 3 days after release to bump Docker
The minimumReleaseAge=3 days will tell renovate to check
the time passed after the release date, and set a pending
status check before that, and a passed status check after
it.

In comnibation with internalChecksFilter="strict", it will
refrain from opening the merge request, untill the time has
passed.

Ref: https://docs.renovatebot.com/configuration-options/#minimumreleaseage

Changelog: changed