Delay docker image updates in GitLab Agent
This change set delays updates to the docker image referenced in the
GitLab pipeline of the GitLab Agent project by 1 day.

The reason for this is that for some unknown reason it takes almost a
day for docker to release all their image variants - especially the
linux/arm64 takes a while.

An example of this behavior can be observed here:
https://gitlab.com/gitlab-org/cluster-integration/gitlab-agent/-/merge_requests/1632