Fix and refactor distribution labels
Our lable assignment was not consistent across our projects. This commit
extracts the Distribution project labels to a shared constant, so
that we don't forget about adding a certain label to a certain project.

Projects can still have their own specific labels if we need it in the
future, as the array can be extended per project.